requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
geographiclib==2.0
geopy==2.4.1
idna==3.7
MarkupSafe==2.1.5
netifaces==0.11.0
prettytable==3.10.2
psutil==6.0.0
speedtest-cli==2.1.3
urllib3==2.2.2
requests==2.32.3
